+++ credly_integration/apps.py
@@ -7,7 +7,7 @@ from __future__ import unicode_literals
 from django.apps import AppConfig
 
 from openedx.core.djangoapps.plugins.constants import (
-    ProjectType, SettingsType, PluginSettings
+    ProjectType, SettingsType, PluginURLs, PluginSettings
 )
 
 
@@ -23,12 +23,28 @@ class CredlyIntegraionConfig(AppConfig):
 
     # Class attribute that configures and enables this app as a Plugin App.
     plugin_app = {
+        PluginURLs.CONFIG: {
+            ProjectType.LMS: {
+                PluginURLs.NAMESPACE: EXTENSIONS_APP_NAME,
+                PluginURLs.APP_NAME: EXTENSIONS_APP_NAME,
+                PluginURLs.REGEX: r'^credly-integration',
+                PluginURLs.RELATIVE_PATH: 'urls',
+            },
+            ProjectType.CMS: {
+                PluginURLs.NAMESPACE: EXTENSIONS_APP_NAME,
+                PluginURLs.APP_NAME: EXTENSIONS_APP_NAME,
+                PluginURLs.REGEX: r'^credly-integration',
+                PluginURLs.RELATIVE_PATH: 'urls',
+            }
+        },
         PluginSettings.CONFIG: {
             ProjectType.CMS: {
                 SettingsType.COMMON: {
                     PluginSettings.RELATIVE_PATH: 'settings.common',
                 },
-                SettingsType.PRODUCTION: {
+            },
+            ProjectType.LMS: {
+                SettingsType.COMMON: {
                     PluginSettings.RELATIVE_PATH: 'settings.common',
                 },
             },