+++ credly_integration/urls.py
@@ -1,0 +1,12 @@
+"""Ubicquia extensions urls."""
+from django.conf.urls import url
+from .views import CredlyReportView
+
+
+urlpatterns = [
+    url(
+        r'^/download_report/$',
+        CredlyReportView.as_view(),
+        name='download_report'
+    ),
+]