+++ user_extensions/apps.py
@@ -1,0 +1,38 @@
+"""
+App configuration for user extensions.
+"""
+
+from __future__ import unicode_literals
+
+from django.apps import AppConfig
+
+from openedx.core.djangoapps.plugins.constants import (
+    ProjectType, SettingsType, PluginSettings
+)
+
+
+EXTENSIONS_APP_NAME = 'user_extensions'
+
+
+class UserExtensionsConfig(AppConfig):
+    """
+    Nasa Tops user extensions configuration.
+    """
+    name = EXTENSIONS_APP_NAME
+    verbose_name = 'User Extensions'
+
+    # Class attribute that configures and enables this app as a Plugin App.
+    plugin_app = {
+        PluginSettings.CONFIG: {
+            ProjectType.CMS: {
+                SettingsType.COMMON: {
+                    PluginSettings.RELATIVE_PATH: 'settings.common',
+                },
+            },
+            ProjectType.LMS: {
+                SettingsType.COMMON: {
+                    PluginSettings.RELATIVE_PATH: 'settings.common',
+                },
+            },
+        }
+    }